refactor : MonthView Component Extraction
